third_party/WebKit/LayoutTests/shadow-dom/css-focus-pseudo-match-shadow-host5-expected.txt

testing nested distribution
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
#input1, #input2 are (re)distributed in the same treescope as #input3, but :focus on shadow host only matches when a focused element is under its shadow tree.
(1/4) top and inner shadow do not delegate focus.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
(2/4) top shadow delegates, but inner shadow does not.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
(3/4) top shadow does not delegate, but inner shadow does.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
(4/4) both shadow hosts delagate focus.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE
